user_management/requirements.txt
# Web framework and API
fastapi[standard]
uvicorn

# Data validation and modeling
pydantic
pydantic_core

# MongoDB ORM
mongoengine
pymongo

# Session and token handling
itsdangerous
werkzeug
PyJWT

# Typing and core tools
typing_extensions

# Testing
pytest
pytest-mock
mongomock

# Google loginn
python-dotenv
httpx
google-auth
google-auth-oauthlib
requests